Added files via upload
Got the add entry page working how I wanted it to with the added bonus of having the script set the action of the form to different values depending on the entry you're adding. So no worrying about if/else statements involving team/match variables when you're adding a league, because it jumps to the very specific submit_league.php file.
Or at least it should. Let's not count eggs until I get it working with Database info.