Game Clock *In Progress*
I was unable to test, if the clock actually works in the app. I could
not get past registration or login page due to lack of knowledge on how
to.  Clock is just operational client side, nothing server side yet. I
am still working to get the server side of the clock established.