Client side connects to server again
Our free trial for the google cloud server expired but I created a new
free trial which kept all of our saved files. The external ip address to
the virtual machine was the only thing that changed. I changed the
address on the client side to match the new ip address.